updates to automation.
rivnet may be a lot cause due to processing times and supervisory choices for extents. Alterantive would be download a huge area up front and then use that as the DEM, but that will eat RAM.

streamstats API is now totally automated except for the last few lines, which require a visual check for sheds with massive holes, angular shapes, etc.